Combine some too small files, ensure .npmrc is ignored
