--- src/APIClientConfiguration.ts
@@ -1,4 +1,0 @@
-export interface APIClientConfiguration {
-  baseURL: string
-  apiKey: string
-}
