--- src/Application.ts
@@ -1,38 +1,0 @@
-export interface Application {
-  id: number
-  guid: string
-  accessType: string
-  connectionTimeout?: number
-  readTimeout?: number
-  initTimeout?: number
-  idleTimeout?: number
-  maxProcesses?: number
-  minProcesses?: number
-  maxConnsPerProcess?: number
-  loadFactor?: number
-  url: string
-  vanityUrl: boolean
-  name: string
-  title?: string
-  bundleId: number
-  appMode: number
-  contentCategory: string
-  hasParameters: boolean
-  createdTime: Date
-  lastDeployedTime: Date
-  rVersion?: string
-  pyVersion?: string
-  buildStatus: number
-  runAs?: string
-  runAsCurrentUser: boolean
-  description: string
-  appRole: string
-  ownerFirstName: string
-  ownerLastName: string
-  ownerUsername: string
-  ownerGuid: string
-  ownerEmail: string
-  ownerLocked: boolean
-  isScheduled: boolean
-  git?: string
-}
