--- src/DeployTaskResponse.ts
@@ -1,7 +1,0 @@
-export interface DeployTaskResponse {
-  taskId: string
-  appId: number
-  appGuid: string
-  appUrl: string
-  title: string
-}
