--- src/ExtendedBundleResponse.ts
@@ -1,11 +1,0 @@
-export interface ExtendedBundleResponse {
-  id: number
-  appId: number
-  createdTime: Date
-  updatedTime: Date
-  rVersion?: string
-  pyVersion?: string
-  buildStatus: number
-  size?: number
-  active: boolean
-}
