--- src/ListApplicationParams.ts
@@ -1,5 +1,0 @@
-export interface ListApplicationsParams {
-  count: number
-  start: number
-  cont: string
-}
