--- src/ListApplicationsResponse.ts
@@ -1,8 +1,0 @@
-import { Application } from './Application'
-
-export interface ListApplicationsResponse {
-  applications: Application[]
-  count: number
-  total: number
-  continuation: string
-}
