--- src/ManifestFile.ts
@@ -1,3 +1,0 @@
-export interface ManifestFile {
-  checksum: string
-}
