--- src/NoArgsReturnsNullString.ts
@@ -1,1 +1,0 @@
-export type NoArgsReturnsNullString = () => string | null
