--- src/NoArgsReturnsStringMap.ts
@@ -1,1 +1,0 @@
-export type NoArgsReturnsStringMap = () => Map<string, string>
