--- src/StringReturnsNullString.ts
@@ -1,1 +1,0 @@
-export type StringReturnsNullString = (s: string) => string | null
