--- src/VanityRecordResponse.ts
@@ -1,7 +1,0 @@
-export interface VanityRecordResponse {
-  id?: number
-  appId: number
-  appGuid: string
-  pathPrefix: string
-  createdTime?: Date
-}
