+++ src/function-types.ts
@@ -1,0 +1,7 @@
+export type NoArgsReturnsNullString = () => string | null
+
+export type NoArgsReturnsStringMap = () => Map<string, string>
+
+export type NoArgsReturnsString = () => string
+
+export type StringReturnsNullString = (s: string) => string | null
