+++ src/main.ts
@@ -1,6 +1,5 @@
-export { APIClient } from './APIClient'
-export { APIClientConfiguration } from './APIClientConfiguration'
+export { APIClient, APIClientConfiguration } from './APIClient'
 export { ClientTaskPoller } from './ClientTaskPoller'
-export { DeployTaskResponse } from './DeployTaskResponse'
+export { DeployTaskResponse } from './api-types'
 export { Deployer } from './Deployer'
 export { ListApplicationsPager } from './ListApplicationsPager'
